[DOCS] Minor fixes
