+++ .gitignore
@@ -1,2 +1,4 @@
+bin
+cmake-bin
 antlr4
 test.cpp
