+++ compiler/Driver/TranslationUnit.h
@@ -1,0 +1,14 @@
+#ifndef TS_UNIT_H
+#define TS_UNIT_H
+
+#include <string>
+
+namespace ctc::driver
+{
+    struct translation_unit
+    {
+        std::string source, path;
+    };    
+}
+
+#endif
